Extract Assets test in dedicated spec file
